Use eslint, and refactor typings
