+++ package.json
@@ -5,18 +5,17 @@
   "main": "aedes.js",
   "types": "aedes.d.ts",
   "scripts": {
-    "lint-fix": "standard --fix",
     "lint": "npm run lint:standard && npm run lint:typescript && npm run lint:markdown",
+    "lint:fix": "standard --fix",
     "lint:standard": "standard --verbose | snazzy",
-    "lint:typescript": "standard --parser @typescript-eslint/parser --plugin @typescript-eslint/eslint-plugin test/types/*.ts aedes.d.ts",
+    "lint:typescript": "eslint -c types/.eslintrc.json aedes.d.ts types/**/*.ts test/types/**/*.test-d.ts",
     "lint:markdown": "markdownlint docs/*.md README.md",
+    "test": "npm run lint && npm run unit && npm run test:typescript",
+    "test:ci": "npm run lint && npm run unit -- --cov --coverage-report=lcovonly && npm run test:typescript",
+    "test:report": "npm run lint && npm run unit:report && npm run test:typescript",
+    "test:typescript": "tsd",
     "unit": "tap --no-esm -J test/*.js",
     "unit:report": "tap --no-esm -J test/*.js --cov --coverage-report=html --coverage-report=cobertura | tee out.tap",
-    "typescript": "tsc --project ./test/types/tsconfig.json",
-    "test:report": "npm run lint && npm run unit:report && npm run typescript",
-    "test": "npm run lint && npm run unit && npm run typescript && tsd",
-    "test:types": "tsd",
-    "test:ci": "npm run lint && npm run unit -- --cov --coverage-report=lcovonly && npm run typescript",
     "license-checker": "license-checker --production --onlyAllow=\"MIT;ISC;BSD-3-Clause;BSD-2-Clause\"",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it --disable-metrics"
   },
@@ -41,6 +40,12 @@
   ],
   "tsd": {
     "directory": "test/types"
+  },
+  "standard": {
+    "ignore": [
+      "types/*",
+      "test/types/*"
+    ]
   },
   "repository": {
     "type": "git",
@@ -91,40 +96,40 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@sinonjs/fake-timers": "^7.0.0",
-    "@types/node": "^14.14.10",
-    "@typescript-eslint/eslint-plugin": "^4.9.0",
-    "@typescript-eslint/parser": "^4.9.0",
+    "@sinonjs/fake-timers": "^7.0.2",
+    "@types/node": "^14.14.34",
+    "@typescript-eslint/eslint-plugin": "^4.17.0",
+    "@typescript-eslint/parser": "^4.17.0",
     "concat-stream": "^2.0.0",
     "duplexify": "^4.1.1",
     "license-checker": "^25.0.1",
     "markdownlint-cli": "^0.27.1",
     "mqtt": "^4.2.6",
-    "mqtt-connection": "^4.0.0",
+    "mqtt-connection": "^4.1.0",
     "pre-commit": "^1.2.2",
     "proxyquire": "^2.1.3",
-    "release-it": "^14.2.2",
+    "release-it": "^14.4.1",
     "snazzy": "^9.0.0",
     "standard": "^16.0.3",
     "tap": "^14.11.0",
     "tsd": "^0.14.0",
-    "typescript": "^4.1.2",
+    "typescript": "^4.2.3",
     "websocket-stream": "^5.5.2"
   },
   "dependencies": {
     "aedes-packet": "^2.3.1",
-    "aedes-persistence": "^8.1.1",
+    "aedes-persistence": "^8.1.3",
     "bulk-write-stream": "^2.0.1",
     "end-of-stream": "^1.4.4",
     "fastfall": "^1.5.1",
     "fastparallel": "^2.4.0",
     "fastseries": "^2.0.0",
-    "hyperid": "^2.0.5",
-    "mqemitter": "^4.4.0",
+    "hyperid": "^2.1.0",
+    "mqemitter": "^4.4.1",
     "mqtt-packet": "^6.9.0",
     "readable-stream": "^3.6.0",
     "retimer": "^3.0.0",
     "reusify": "^1.0.4",
-    "uuid": "^8.3.1"
+    "uuid": "^8.3.2"
   }
 }
